bugfix: tx_context_id column values got lost after id bump

+++ src/block.rs
@@ -19,7 +19,7 @@ impl Block {
 #[derive(Default, Debug, Clone, PartialEq, serde_derive::Serialize, serde_derive::Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct Header {
-    pub level: i64,
+    pub level: u32,
     pub proto: i64,
     pub predecessor: String,
     pub timestamp: String,
